# openapi/briefcased.yaml
openapi: 3.1.0
info:
  title: Credential Briefcase Daemon API
  version: 0.1.4
  description: |
    Local admin API for `briefcased`.

    Security model:
    - `/health` is unauthenticated.
    - all `/v1/*` endpoints require `Authorization: Bearer <daemon_token>`.
servers:
  - url: http://127.0.0.1:0
    description: Local daemon (TCP mode, ephemeral port)
components:
  securitySchemes:
    bearerAuth:
      type: http
      scheme: bearer
  schemas:
    ErrorResponse:
      type: object
      required: [code, message]
      properties:
        code:
          type: string
        message:
          type: string
...
    ApproveResponse:
      type: object
      required: [approval_id, approval_token]
      properties:
        approval_id:
          type: string
          format: uuid
        approval_token:
          type: string
...
paths:
...
  /v1/approvals/{id}/approve:
    post:
      summary: Approve a pending request (returns approval token)
      operationId: approve
      security:
        - bearerAuth: []
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: string
            format: uuid
      responses:
        "200":
          description: ok
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ApproveResponse"
        "400":
          description: bad request
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"
        "401":
          description: unauthorized
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"
...